Consolidate view-tier extensions into one place
