+++ Eurofurence/Extensions/NSAttributedString+trim.swift
