+++ Eurofurence/Extensions/NSMutableAttributedString+trim.swift
